RiskHalo_Certification_Challenge/pyproject.toml
[project]
name = "riskhalo-certification-challenge"
version = "0.1.0"
description = "RiskHalo - A Behavioral Risk Intelligence Engine for Intraday Traders"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jupyter>=1.1.1",
    "pandas>=2.0.0",
    "langchain>=1.0.0",
    "langchain-openai>=0.3.7",
    "langchain-community>=0.0.10",
    "langchain-tavily>=0.2.17",
    "langsmith>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "openpyxl>=3.1.0",
    "ipykernel>=6.30.1",
    "chromadb>=0.4.0",
    "ragas==0.2.10",
    "langgraph>=0.3.0",
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "pyyaml",
    "tabulate",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
